+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Changelog
 Version 0.3.6.0
 ---------------
 
-*   Add `IsList` instances for `-XOverloadedLists`.
+*   Add `IsList` instances for `-XOverloadedLists` (@tbidne).
 *   Add `deleteMaybe` for `NEMap`, `NEIntMap`, `NESet`, and `NEIntSet`.
 *   Add indexed traversal instances for `NEMap`, `NEIntMap`, and `NESeq`.
+*   Add `IsNonEmptyList` for uniform non-empty list conversions.
 *   Add union functions for combining possibly-empty and non-empty maps and sets.
 *   Fix `NESeq.adjust'` to use the strict `Seq.adjust'` in the tail.